requirements.txt
opencv_python==4.4.0.46
pandas==1.2.0
progress==1.5
pytesseract==0.3.7
Pillow==8.1.0
pytest==6.2.1
imagehash==4.2.0
art==5.1
colorama==0.4.4